Switched to solid white background for animated sticker conversion
Gif encoding gets significantly faster with solid background because
all pixels that are same in two consecutive frames are then encoded
as transparent, as opposed to just the ones that are transparent to
begin with.

When showing converted animated stickers in-order, it's probably
more important to convert them as fast as possible.